Refactor bullet
