+++ src/components/Tank/TankSensor.ts
