+++ src/types/TankData.ts
@@ -1,5 +1,5 @@
-import Bullet from '@/components/Bullet';
-import TankSensor from '@/components/TankSensor';
+import Bullet from '@/components/Projectile/Bullet';
+import TankSensor from '@/components/Tank/TankSensor';
 
 interface TankData {
   blocked: {
